powerpc: boot: add simpleboot support for XTM330
While the XTM330 comes with a bootloader supporting device trees some
limitations of the bootloader (maximum gzip decompress size of 8MiB,
upper malloc limit of 16MiB (0x1000000)) make it infeasible to boot a
modern Linux directly. With any modern Kernel the devicetree blob will
end up phyiscally below the kernel in memory since the uncompressed kernel
is always larger than 16 MiB due to strict RWX alignment requirements and
when patching the devicetree the bootloader tries to copy the devicetree
to a phyiscal address < 0x1000000. Since Linux only supports a single TLB
mapping for both kernel and devicetree while also requiring the virtual
address of the mapping to map to _stext it is not possible to setup a
compatible kernel and DTB memory layout from the bootloader.
Wrap kernel and dtb into a simpleboot image instead to work around this
issue.

Signed-off-by: Tobias Schramm <tobias@t-sys.eu>